SW-4895 Create view-agnostic DeliverablesTable from the working contents of the AcceleratorDeliverablesList
